+++ changelog.md
@@ -2,6 +2,8 @@
 
 **What's New**
 * [InputAddon]: added InputAddon component that can be used for Prefix/Suffix with ControlGroup.
+* [Badge]: deprecated mode fill="transparent", and deprecated mode fill="semitransparent", they will be removed in future release.
+* [Badge]: removed mode fill="white", and "none", use fill="outline" + color instead them.
 * [useVirtualList]: `scrollTo` and `scrollToIndex` API was changed. `scrollTo` is an object, with index, behavior and align configuration.
 * [DataTable]: added pinned rows functionality.